+++ crates/findex/src/ffi/core/mod.rs
@@ -14,6 +14,7 @@ pub const NUMBER_OF_ENTRY_TABLE_LINE_IN_BATCH: usize = 100;
 
 /// Implements Findex traits.
 pub struct FindexUser {
+    pub(crate) entry_table_number: usize,
     pub(crate) progress: Option<ProgressCallback>,
     pub(crate) fetch_all_entry_table_uids: Option<FetchAllEntryTableUidsCallback>,
     pub(crate) fetch_entry: Option<FetchEntryTableCallback>,