+++ crates/findex/src/wasm_bindgen/core/utils.rs
@@ -50,7 +50,7 @@ pub async fn fetch_uids<const UID_LENGTH: usize>(
     uids: &HashSet<Uid<UID_LENGTH>>,
     fetch_callback: &Fetch,
     source_for_errors: &'static str,
-) -> Result<EncryptedTable<UID_LENGTH>, FindexWasmError> {
+) -> Result<Vec<(Uid<UID_LENGTH>, Vec<u8>)>, FindexWasmError> {
     // Convert Inputs to array of Uint8Array
     let input = Array::new();
     for uid in uids {
@@ -61,8 +61,8 @@ pub async fn fetch_uids<const UID_LENGTH: usize>(
     // perform the call
     let output = callback!(fetch_callback, input);
 
-    // parse results into HashMap
-    js_value_to_encrypted_table(&output, source_for_errors).map_err(FindexWasmError::from)
+    // parse results into Vec
+    js_value_to_entry_table_items(&output, source_for_errors).map_err(FindexWasmError::from)
 }
 
 #[inline]
@@ -78,10 +78,10 @@ pub fn set_bytes_in_object_property(
     })
 }
 
-pub fn js_value_to_encrypted_table<const UID_LENGTH: usize>(
+pub fn js_value_to_entry_table_items<const UID_LENGTH: usize>(
     encrypted_table: &JsValue,
     callback_name_for_errors: &'static str,
-) -> Result<EncryptedTable<UID_LENGTH>, JsValue> {
+) -> Result<Vec<(Uid<UID_LENGTH>, Vec<u8>)>, JsValue> {
     if !Array::is_array(encrypted_table) {
         return Err(JsValue::from_str(
             format!(
@@ -96,7 +96,7 @@ pub fn js_value_to_encrypted_table<const UID_LENGTH: usize>(
     }
 
     let array = Array::from(encrypted_table);
-    let mut encrypted_table = EncryptedTable::<UID_LENGTH>::with_capacity(array.length() as usize);
+    let mut encrypted_table = Vec::with_capacity(array.length() as usize);
     let object_source_for_errors =
         ObjectSourceForErrors::ReturnedFromCallback(callback_name_for_errors);
     for (i, try_obj) in array.values().into_iter().enumerate() {
@@ -117,7 +117,7 @@ pub fn js_value_to_encrypted_table<const UID_LENGTH: usize>(
         let uid = get_bytes_from_object_property(&obj, "uid", &object_source_for_errors, i)?;
 
         let value = get_bytes_from_object_property(&obj, "value", &object_source_for_errors, i)?;
-        encrypted_table.insert(
+        encrypted_table.push((
             Uid::try_from_bytes(&uid).map_err(|e| {
                 JsValue::from_str(
                     format!(
@@ -129,7 +129,7 @@ pub fn js_value_to_encrypted_table<const UID_LENGTH: usize>(
                 )
             })?,
             value.clone(),
-        );
+        ));
     }
     Ok(encrypted_table)
 }